Implement root adapter and link original adapter to it
