+++ app/src/main/java/com/github/palFinderTeam/palfinder/ProfileViewModel.kt
@@ -6,6 +6,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import androidx.fragment.app.activityViewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.github.palFinderTeam.palfinder.meetups.FirebaseMeetUpService
 import com.github.palFinderTeam.palfinder.meetups.MeetUp
@@ -16,7 +17,9 @@ import com.github.palFinderTeam.palfinder.profile.ProfileService
 import com.github.palFinderTeam.palfinder.profile.ProfileUser
 import com.github.palFinderTeam.palfinder.utils.Response
 import com.github.palFinderTeam.palfinder.utils.SearchedFilter
+import dagger.hilt.android.AndroidEntryPoint
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
@@ -25,10 +28,9 @@ import javax.inject.Inject
 @HiltViewModel
 class ProfileViewModel @Inject constructor(
     private val profileService: ProfileService,
-    private val meetUpService: MeetUpRepository
+    //private val meetUpService: MeetUpRepository
 ) : ViewModel() {
 
-    val viewModel: MapListViewModel by activityViewModels()
 
     private val _profile: MutableLiveData<Response<ProfileUser>> = MutableLiveData()
     val profile: LiveData<Response<ProfileUser>> = _profile
@@ -56,39 +58,39 @@ class ProfileViewModel @Inject constructor(
         }
     }
 
-    /**
-     * Get the list of meetups of a user
-     * @param userId The ID of a user
-     */
-    fun fetchUserMeetups(userId: String){
-        viewModelScope.launch {
-
-            // TODO: change to user list when Zac implements it
-            meetUpService.getAllMeetUpsResponse().collect{
-
-                if (it is Response.Success) {
-                    val userMeetups = it.data
-
-                    adapter = MeetupListAdapter(
-                        userMeetups, userMeetups.toMutableList(),
-                        SearchedFilter(
-                            userMeetups, userMeetups.toMutableList(), ::filterTags
-                        ) {
-                            adapter.notifyDataSetChanged()
-                        },
-                        viewModel.searchLocation.value!!
-                    )
-
-                    _listOfMeetUpResponse.postValue(it)
-                }
-
-            }
-        }
-    }
-
-    private fun filterTags(meetup: MeetUp): Boolean {
-        return true
-    }
+//    /**
+//     * Get the list of meetups of a user
+//     * @param userId The ID of a user
+//     */
+//    fun fetchUserMeetups(userId: String){
+//        viewModelScope.launch {
+//
+//            // TODO: change to user list when Zac implements it
+//            meetUpService.getAllMeetUpsResponse().collect{
+//
+//                if (it is Response.Success) {
+//                    val userMeetups = it.data
+//
+//                    adapter = MeetupListAdapter(
+//                        userMeetups, userMeetups.toMutableList(),
+//                        SearchedFilter(
+//                            userMeetups, userMeetups.toMutableList(), ::filterTags
+//                        ) {
+//                            adapter.notifyDataSetChanged()
+//                        },
+//                        viewModel.searchLocation.value!!
+//                    )
+//
+//                    _listOfMeetUpResponse.postValue(it)
+//                }
+//
+//            }
+//        }
+//    }
+//
+//    private fun filterTags(meetup: MeetUp): Boolean {
+//        return true
+//    }
 
 
 }
